+++ src/main/java/core/basesyntax/Rectangle.java
@@ -1,10 +1,10 @@
 package core.basesyntax;
 
-public class Rectangle extends Figure{
-    private double width;
-    private double length;
+public class Rectangle extends Figure {
+    private int width;
+    private int length;
 
-    public Rectangle(double width, double length, String color) {
+    public Rectangle(int width, int length, String color) {
         super(color);
         this.width = width;
         this.length = length;
@@ -12,13 +12,13 @@ public class Rectangle extends Figure{
 
     @Override
     public double area() {
-        double area =  width * length;
+        double area = width * length;
         return area;
     }
 
     @Override
     public void draw() {
-        System.out.println("Figure: rectangle, area: " + area() + " sq.units, " +
-                "width: " + width + " length: " + length + ", color: " + getColor());
+        System.out.println("Figure: rectangle, area: " + area() + " sq.units, "
+                + "width: " + width + " length: " + length + ", color: " + getColor());
     }
 }
